+++ untitled/src/Venta.java
@@ -1,3 +1,3 @@
 public interface Venta {
-    public float montoVenta();
+    public double montoVenta();
 }
